move connections context to common directory + load connections provider in open gov window
