+++ src/renderer/contexts/common/Connections/types.ts
